Repository CRUD

+++ app/src/main/java/com/brahimboulkriaat/psa/model/City.kt
@@ -1,3 +1,3 @@
 package com.brahimboulkriaat.psa.model
 
-data class City(val id: Long, val name: String, val lat: Double, val lon: Double)
+data class City(val id: Int, val name: String, val lat: Double, val lon: Double)
